Update makefile (#134)
Added a compiler macro to make it easier to use something other than gcc. Added some macros to make the targets more readable (I think). And tries to detect whether in a windows environment or not in order to make pgo work. Wouldn't be surprised if there are cases where it fails, but it correctly differentiates cmd and git bash on my windows machine and pgo works in both.