Defer allocation of TT until necessary (#151)
This change avoids allocating the TT with the default size only to free it and reallocate in the likely event that a hash size is specified by the user. This speeds up initialization in most cases.

If told to search before getting an 'isready' weiss will now crash. Checking for allocated TT on every 'go' showed a 4 elo loss so I will allow this behavior. With that said, the uci protocol specifies an isready must be given before starting a search to allow initialization to finish, and recommends it after any command which could take some time. Reasonable interpretation is that isready should always be given between setting hash size and starting a search.

ELO   | 2.03 +- 4.25 (95%)
SPRT  | 10.0+0.1s Threads=1 Hash=32MB
LLR   | 2.99 (-2.94, 2.94) [-4.00, 1.00]
Games | N: 15751 W: 4898 L: 4806 D: 6047
http://chess.grantnet.us/viewTest/4456/