Retire dev cli (#155)
The dev cli was inherited from VICE and was used to play games in simple way, as well as run tests. The game playing feature was removed in #133 leaving only the tests. This patch moves the tests to the uci loop, still avaiable only in dev compilations. Also added conditional groups for dev/debug functions.

No functional change.