Fix printing size_t on 32bit OS (#159)
On systems where size_t is not 64 bits, master produces warnings and prints nonsense sizes and entry counts for the transposition table. This patch fixes this by using %zu to print size_t on non-windows systems. On windows, MingW doesn't support %zu as it uses MSVCRT which conforms to C89 while the 'z' length modifier was added in C99.

No functional change.