Speedup en passant square calculation (#164)
Xor'ing a square on rank 3 or 4 swaps it to the other one, and similarly for 5 and 6. Given the destination square of a pawn start we can calculate the en passant square this way. We can also calculate where the pawn we captured with an en passant move was. This is much cleaner, and faster, than the old method.

ELO   | 6.72 +- 6.02 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=32MB
LLR   | 3.07 (-2.94, 2.94) [-2.50, 2.50]
Games | N: 8062 W: 2621 L: 2465 D: 2976
http://chess.grantnet.us/viewTest/4528/