Request huge pages on linux (#192)
Based on a recent patch for Ethereal by @skiminki , this patch aligns the transposition table to 2MB and requests the use of huge pages on linux. When huge pages are available and not used automatically this is potentially a big speedup. When huge pages are already used, the alignment will likely give a smaller speedup.

No functional change.

ELO | 4.77 +- 4.73 (95%)
SPRT | 8.0+0.08s Threads=1 Hash=32MB
LLR | 2.96 (-2.94, 2.94) [-3.00, 1.00]
Games | N: 10936 W: 2966 L: 2816 D: 5154
http://chess.grantnet.us/viewTest/4790/