Cleanup (#197)
Clean up some code, add some comments, and rename some variables.

I decided to retire the dev-mode matetest as with the greatly increased pruning since its inception there is no longer any guarantee to find even fairly short mates. The test would have to be redesigned completely to be useful again. Rather than doing this I think it would just be better to make it as a standalone python script or such. This would as an added bonus allow it to test other engines as well.

No functional change.