Remove test suites (#211)
Test suites moved to a separate repo to reduce the size of this one. Only one epd file is currently used by weiss; all.epd for ensuring the eval is symmetric. To use this test download the EPDs folder from https://github.com/TerjeKir/Chess-test-suites and place it in the weiss folder (next to /src).

No functional change.