Probe Syzygy DTZ in root (#258)
Probe Syzygy DTZ tablebases and use the recommended move in root when made to search a position found in the tablebases. This means Weiss won't mess up a win or draw, but it might not play the best defense when losing, choosing the slowest, but possibly trivially easy, loss over the most difficult.