Graceful exit 2 (#289)
The last patch fixed the case when Weiss is killed by sending it 'quit', but it did not fix segfaulting when killed by ctrl+c. This destructor is more hassle than being nice and freeing allocated memory on exit is worth - any modern OS will free everything when the program exits anyway so I will just remove it.