Simplify time management (#303)
Currently game ply is reset to 0 whenever the 50 move rule resets, so the ply part of the time management which was intended to increase time usage as the game goes on, rather increases time usage the further we are from the last 50mr reset. 

Removing it shows no significant loss at normal time controls.