Resolve an issue where typically main threads were never pthread_join()'ed
Over time, this issue led to a crash. On average, after about ~1800 games played on 8 Threads with 10+.01s time control, Weiss would crash due to resource issues caused by the hanging pthreads. This patch detaches the main search thread so that we can trust it to clean up its own resources.

A self-play test of ~5600 games with the same settings as above was done to confirm the fix.

This same patch was applied to Ethereal to address the very same issue.