Count root moves before search (#519)
It seems the time limit for only-moves was sometimes applied even when there were multiple root moves at TCEC. As seen in SuFi testing games for season 21. I have no idea how that could happen, and I was unable to reproduce the behavior.

To avoid this just count root moves before starting the search instead of relying on number of moves searched in root.