Fix bug with MultiPV (#539)
When there were fewer legal (or restricted by searchmoves) moves than MultiPV and the scores of the searched moves were negative they would be sorted behind empty moves, and skipped in printing. This patch includes the searchmoves restriction when counting rootmoves, and sorts at most as many PVs as there are rootmoves.