Validate move from TT earlier (#548)
The pseudo-legality check is now cheaper than before, so the cost of validating the tt move earlier is low enough that it doesn't seem to matter. Validating the tt move earlier guards against hash collisions as the hash would have to match as well as the move be pseudo-legal for the wrong info to be used. Same for mismatched entries caused by data races when using multiple threads where parts of the entry are from one position while the rest is from another. I doubt this matters for playing strength, but maybe in conditions like at TCEC it makes a slight difference.